Move the decorators to decorators.py
